+++ Rocksmith2014Xml/IHasTimeCode.cs
@@ -1,7 +1,13 @@
 ﻿namespace Rocksmith2014Xml
 {
+    /// <summary>
+    /// Interface for classes that contain a time code in milliseconds.
+    /// </summary>
     public interface IHasTimeCode
     {
-        float Time { get; }
+        /// <summary>
+        /// Time in milliseconds.
+        /// </summary>
+        uint Time { get; }
     }
 }
